merge system and user nodepools subnets for the sake of simplicity
System + User Nodepool will co-exist in the same /22 subnet.
That means our spoke will return to three subnets instead of four. This
is based on the following idea that this is a single-tenant cluster.

Multi-tenant or other advanced workloads may have nodepool isolation requirements
that might demand more (and likely smaller) subnets.