redude ILB address space to /28
Ingress Subnet will move to a /28 – this is more than we need for the story we want to tell,
But this will align with a better “minimum” for future motorcycle stories we want to tell.

Generally speaking, we want customers to opt for one ingress controller per cluster for simplicity
to start with (when single-tenant) vs ingress controller namespace, etc.

Multi-tenant, multiple SSL termination rules, single PPE
supporting dev/QA/UAT, etc could lead to needing more
ingress controllers, but for baseline, we should start with one.